Make foobarplugin part of the test only
