--- testing/foobar/setup.py
@@ -1,6 +1,0 @@
-from setuptools import setup
-
-setup(
-    py_modules=["foobarplugin"],
-    entry_points={"pytest11": ["foobarplugin = foobarplugin"]},
-)
