+++ tox.ini
@@ -11,7 +11,6 @@ deps =
   pytestlatest: pytest
   pytestmaster: git+https://github.com/pytest-dev/pytest.git@master
   pytestfeatures: git+https://github.com/pytest-dev/pytest.git@features
-  -etesting/foobar
 commands=
   pytest {posargs}
